Relocate currency configuration
